Make glGetQueryObjectuiv(GL_QUERY_RESULT_AVAILABLE) wait for the available result
This patch is designed to fix the problem with the deqp occlusion test
on virgl mesa implementation.  The test goes like this:

```
	glGenQueries    (1, &queryIds[0]);
	glBeginQuery    (m_queryMode, queryIds[0]);

	// draw polygons

	glEndQuery      (m_queryMode);
	glFinish        ();

	glGetQueryObjectuiv(queryIds[0], GL_QUERY_RESULT_AVAILABLE, &resultAvailable);
	if (resultAvailable == GL_FALSE) {
		// test fails
	}

	glGetQueryObjectuiv(queryIds[0], GL_QUERY_RESULT, &result);
	// use `result` to find out if the occlusion test works well
```

The test fails because glFinish() should wait for all the queries are finished
so that glGetQueryObjectuiv(GL_QUERY_RESULT_AVAILABLE) in the next line returns true.
Unfortunately it is not implemented and glFinish() works here as just glFlush() + fence.

In mesa3d implementation there is an internal flag PIPE_FLUSH_HINT_FINISH that
associated in the core of mesa3d with the fence that is created in glFinish().
Unfortunately I did not find any driver that pays attention to that flag.

NB: glFlush() implementation does create internally a fence and waits for it.
So all commands are guaranteed to be completed.  The problem here exists just
with the *queries* that are created by glGenQueries()/glBeginQuery() commands.

So it seems that every driver has the same problem with the implementation, but
it shows up in virgl probably because all the hardware drivers provide
the results of the tests quickly and glGetQueryObjectuiv(GL_QUERY_RESULT_AVAILABLE)
returns true at the first invocation.

The correct fix to this problem would be to handle this flag correctly in the
virgl mesa3d gallium driver.  Unfortunatelly that would also mean that this
flag (or some derivative of it that is internal to the driver) should also
be signalled to the host side, and that in turn requires changing
- mesa3d virgl gallium driver / virgl linux kernel drm interface
- linux virgl drm driver
- VIRTIO GPU interface
- virglrenderer
That is a lot of work.  So instead I changed glGetQueryObjectuiv(GL_QUERY_RESULT_AVAILABLE)
so that it always wait for the result and returns true.

This patch makes changes to mesa3d core.
Another possibility is to make changes to the virgl gallium driver.

Cherry-picked from:
Repository: git@github.com:qcc-auto/qcvirtio-mesa.git
Revision: 58bbc17ee09ee92fd413e870c083574eab4d83f1

Change-Id: I8ce27e17472420231c83608fdac86d606252a035